+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.17.0 (2026-07-15)
+
+
+### Gemini
+
+* **gemini:** add fix-license-scan skill
+
+
+### Features
+
+* add benchmark for ShmSlabPool
+
 ## 0.16.0 (2026-07-10)